Added shared helper
