+++ Vostok.Metrics.System/Host/NativeHostMetricsCollector_Windows.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.ComponentModel;
 using System.Runtime.InteropServices;
 using Vostok.Metrics.System.Helpers;
 
@@ -15,19 +14,6 @@ namespace Vostok.Metrics.System.Host
             CollectMemoryMetrics(metrics);
         }
 
-        private static void ThrowOnError()
-        {
-            var exception = new Win32Exception();
-
-            throw new Win32Exception(exception.ErrorCode, exception.Message + $" Error code = {exception.ErrorCode} (0x{exception.ErrorCode:X}).");
-        }
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        private static extern bool GetSystemTimes(
-            out FILETIME idleTime,
-            out FILETIME kernelTime,
-            out FILETIME userTime);
-
         [DllImport("psapi.dll", SetLastError = true)]
         private static extern bool GetPerformanceInfo(
             [Out] out PERFORMANCE_INFORMATION ppsmemCounters,
@@ -37,8 +23,8 @@ namespace Vostok.Metrics.System.Host
         {
             try
             {
-                if (!GetSystemTimes(out var idleTime, out var systemKernel, out var systemUser))
-                    ThrowOnError();
+                if (!WinMetricsCollectorHelper.GetSystemTimes(out var idleTime, out var systemKernel, out var systemUser))
+                    WinMetricsCollectorHelper.ThrowOnError();
 
                 var systemTime = systemKernel.ToUInt64() + systemUser.ToUInt64();
 
@@ -55,7 +41,7 @@ namespace Vostok.Metrics.System.Host
             try
             {
                 if (!GetPerformanceInfo(out var perfInfo, sizeof(PERFORMANCE_INFORMATION)))
-                    ThrowOnError();
+                    WinMetricsCollectorHelper.ThrowOnError();
 
                 metrics.MemoryAvailable = (long) perfInfo.PhysicalAvailable * (long) perfInfo.PageSize;
                 metrics.MemoryCached = (long) perfInfo.SystemCache * (long) perfInfo.PageSize;
@@ -85,20 +71,6 @@ namespace Vostok.Metrics.System.Host
             public readonly uint HandleCount;
             public readonly uint ProcessCount;
             public readonly uint ThreadCount;
-        }
-
-        [StructLayout(LayoutKind.Sequential, Pack = 1)]
-        private struct FILETIME
-        {
-            public readonly uint dwLowDateTime;
-            public readonly uint dwHighDateTime;
-
-            public ulong ToUInt64()
-            {
-                var high = (ulong) dwHighDateTime;
-                var low = (ulong) dwLowDateTime;
-                return (high << 32) | low;
-            }
         }
     }
 }
